BUILD: inlined common dir and removed excessive clean targets
The common dir is now "multiplexed" into the seperate product
directories (*batch* and *rake*). This makes it easier for the user to
install the product, since he only needs to perform a single copy
action.

Dependency on the clean package of Rake has been removed, since it
unnecessary (the packagetask already provides a target for this).