Fixes garbled output
This changes the implementation to capture the output of a build. Now,
it starts gis.exe with a different COMSPEC, which invokes rubyw.exe with
a placeholder script to invoke the actual BUILD_IMAGE.cmd script. We're
using a slightly modified version of GE (no explicit redirection, less
output, etc.).