Redirects stderr to stdout of builds.
Output to stderr were lost, because that wasn't "captured" with our
IO.popen calls. This resulted in loss of error messages of gis.exe.

This prevents those messages from getting lost.

Still the build doesn't fail if an image is not present, but involves a
bit more work, since we need to capture the stderr stream in a seperate
thread. At least, I think so.